+++ src/views/MenuPublicView.vue
@@ -1,0 +1,7 @@
+<script setup>
+import MenuPublicIndex from '../components/Menu/MenuPublicIndex.vue'
+</script>
+
+<template>
+    <MenuPublicIndex />
+</template>